+++ pallets/foreign-investments/src/lib.rs
@@ -43,27 +43,20 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-use cfg_traits::{
-	investments::{Investment, TrancheCurrency},
-	IdentityCurrencyConversion, PoolInspect,
-};
-use cfg_types::investments::{
-	CollectedAmount, ExecutedForeignCollect, ExecutedForeignDecreaseInvest, Swap,
-};
-use frame_support::{dispatch::DispatchResult, ensure};
+use cfg_types::investments::Swap;
 pub use pallet::*;
 use parity_scale_codec::{Decode, Encode, MaxEncodedLen};
 use scale_info::TypeInfo;
-use sp_runtime::{
-	traits::{EnsureAdd, EnsureAddAssign, EnsureSub, EnsureSubAssign, Saturating, Zero},
-	ArithmeticError, DispatchError,
-};
 
 #[cfg(test)]
 mod mock;
 
 #[cfg(test)]
 mod tests;
+
+mod entities;
+mod impls;
+mod swaps;
 
 #[derive(
 	Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Debug, Encode, Decode, TypeInfo, MaxEncodedLen,
@@ -80,320 +73,17 @@ pub type ForeignId<T> = (
 	Action,
 );
 
-/// Hold the base information of a foreign investment/redemption
-#[derive(Clone, PartialEq, Eq, Debug, Encode, Decode, TypeInfo, MaxEncodedLen)]
-#[scale_info(skip_type_params(T))]
-pub struct BaseInfo<T: Config> {
-	foreign_currency: T::CurrencyId,
-	collected: CollectedAmount<T::Balance>,
-}
-
-impl<T: Config> BaseInfo<T> {
-	fn new(foreign_currency: T::CurrencyId) -> Result<Self, DispatchError> {
-		Ok(Self {
-			foreign_currency,
-			collected: CollectedAmount::default(),
-		})
-	}
-
-	fn ensure_same_foreign(&self, foreign_currency: T::CurrencyId) -> DispatchResult {
-		ensure!(
-			self.foreign_currency == foreign_currency,
-			Error::<T>::MismatchedForeignCurrency
-		);
-
-		Ok(())
-	}
-}
-
-/// Hold the information of a foreign investment
-#[derive(Clone, PartialEq, Eq, Debug, Encode, Decode, TypeInfo, MaxEncodedLen)]
-#[scale_info(skip_type_params(T))]
-pub struct InvestmentInfo<T: Config> {
-	/// General info
-	base: BaseInfo<T>,
-
-	/// Total amount of pool currency increased for this investment
-	increased_pool_amount: T::Balance,
-
-	/// Total swapped amount pending to execute for decreasing the investment.
-	/// Measured in foreign currency
-	decrease_swapped_amount: T::Balance,
-}
-
-impl<T: Config> InvestmentInfo<T> {
-	fn new(foreign_currency: T::CurrencyId) -> Result<Self, DispatchError> {
-		Ok(Self {
-			base: BaseInfo::new(foreign_currency)?,
-			increased_pool_amount: T::Balance::default(),
-			decrease_swapped_amount: T::Balance::default(),
-		})
-	}
-
-	fn pre_increase_swap(
-		&mut self,
-		investment_id: T::InvestmentId,
-		foreign_amount: T::Balance,
-	) -> Result<Swap<T::Balance, T::CurrencyId>, DispatchError> {
-		// NOTE: This line will be removed with market ratios
-		let pool_amount = T::CurrencyConverter::stable_to_stable(
-			pool_currency_of::<T>(investment_id)?,
-			self.base.foreign_currency,
-			foreign_amount,
-		)?;
-
-		self.increased_pool_amount.ensure_add_assign(pool_amount)?;
-
-		Ok(Swap {
-			currency_in: pool_currency_of::<T>(investment_id)?,
-			currency_out: self.base.foreign_currency,
-			amount_in: pool_amount,
-		})
-	}
-
-	/// Decrease an investment taking into account that a previous increment
-	/// could be pending
-	fn pre_decrease_swap(
-		&mut self,
-		who: &T::AccountId,
-		investment_id: T::InvestmentId,
-		foreign_amount: T::Balance,
-	) -> Result<Swap<T::Balance, T::CurrencyId>, DispatchError> {
-		// NOTE: This line will be removed with market ratios
-		let pool_amount = T::CurrencyConverter::stable_to_stable(
-			pool_currency_of::<T>(investment_id)?,
-			self.base.foreign_currency,
-			foreign_amount,
-		)?;
-
-		self.increased_pool_amount
-			.ensure_sub_assign(pool_amount)
-			.map_err(|_| Error::<T>::TooMuchDecrease)?;
-
-		let pool_currency = pool_currency_of::<T>(investment_id)?;
-		let pending_pool_amount_increment =
-			Pallet::<T>::pending_swap_amount(who, investment_id, pool_currency, Action::Investment);
-
-		let decrement = pool_amount.saturating_sub(pending_pool_amount_increment);
-		if !decrement.is_zero() {
-			T::Investment::update_investment(
-				who,
-				investment_id,
-				T::Investment::investment(who, investment_id)?.ensure_sub(decrement)?,
-			)?;
-		}
-
-		Ok(Swap {
-			currency_in: self.base.foreign_currency,
-			currency_out: pool_currency_of::<T>(investment_id)?,
-			amount_in: foreign_amount,
-		})
-	}
-
-	/// Increase an investment taking into account that a previous decrement
-	/// could be pending
-	fn post_increase_swap(
-		&mut self,
-		who: &T::AccountId,
-		investment_id: T::InvestmentId,
-		pool_amount: T::Balance,
-	) -> DispatchResult {
-		self.decrease_swapped_amount = T::Balance::default();
-
-		if !pool_amount.is_zero() {
-			T::Investment::update_investment(
-				&who,
-				investment_id,
-				T::Investment::investment(&who, investment_id)?.ensure_add(pool_amount)?,
-			)?;
-		}
-
-		Ok(())
-	}
-
-	fn post_decrease_swap(
-		&mut self,
-		investment_id: T::InvestmentId,
-		swapped_amount: T::Balance,
-		pending_amount: T::Balance,
-	) -> Result<Option<ExecutedForeignDecreaseInvest<T::Balance, T::CurrencyId>>, DispatchError> {
-		self.decrease_swapped_amount
-			.ensure_add_assign(swapped_amount)?;
-
-		if pending_amount.is_zero() {
-			// NOTE: How make this works with market ratios?
-			let remaining_foreign_amount = T::CurrencyConverter::stable_to_stable(
-				self.base.foreign_currency,
-				pool_currency_of::<T>(investment_id)?,
-				self.remaining_pool_amount()?,
-			)?;
-
-			let msg = ExecutedForeignDecreaseInvest {
-				amount_decreased: self.decrease_swapped_amount,
-				foreign_currency: self.base.foreign_currency,
-				amount_remaining: remaining_foreign_amount,
-			};
-
-			self.decrease_swapped_amount = T::Balance::default();
-
-			return Ok(Some(msg));
-		}
-
-		Ok(None)
-	}
-
-	fn post_collect(
-		&mut self,
-		investment_id: T::InvestmentId,
-		collected: CollectedAmount<T::Balance>,
-	) -> Result<ExecutedForeignCollect<T::Balance, T::CurrencyId>, DispatchError> {
-		self.base.collected.increase(&collected)?;
-
-		// NOTE: How make this works with market ratios?
-		let remaining_foreign_amount = T::CurrencyConverter::stable_to_stable(
-			self.base.foreign_currency,
-			pool_currency_of::<T>(investment_id)?,
-			self.remaining_pool_amount()?,
-		)?;
-
-		// NOTE: How make this works with market ratios?
-		let collected_foreign_amount = T::CurrencyConverter::stable_to_stable(
-			self.base.foreign_currency,
-			pool_currency_of::<T>(investment_id)?,
-			collected.amount_payment,
-		)?;
-
-		let msg = ExecutedForeignCollect {
-			currency: self.base.foreign_currency,
-			amount_currency_payout: collected_foreign_amount,
-			amount_tranche_tokens_payout: collected.amount_collected,
-			amount_remaining: remaining_foreign_amount,
-		};
-
-		Ok(msg)
-	}
-
-	fn is_completed(&self) -> Result<bool, DispatchError> {
-		Ok(self.decrease_swapped_amount.is_zero() && self.remaining_pool_amount()?.is_zero())
-	}
-
-	fn remaining_pool_amount(&self) -> Result<T::Balance, ArithmeticError> {
-		self.increased_pool_amount
-			.ensure_sub(self.base.collected.amount_payment)
-	}
-}
-
-/// Hold the information of an foreign redemption
-#[derive(Clone, PartialEq, Eq, Debug, Encode, Decode, TypeInfo, MaxEncodedLen)]
-#[scale_info(skip_type_params(T))]
-pub struct RedemptionInfo<T: Config> {
-	/// General info
-	base: BaseInfo<T>,
-
-	/// Amount of tranche tokens pending to redeem
-	pending_tranche_tokens: T::Balance,
-
-	/// Total swapped amount pending to execute.
-	swapped_amount: T::Balance,
-}
-
-impl<T: Config> RedemptionInfo<T> {
-	fn new(foreign_currency: T::CurrencyId) -> Result<Self, DispatchError> {
-		Ok(Self {
-			base: BaseInfo::new(foreign_currency)?,
-			pending_tranche_tokens: T::Balance::default(),
-			swapped_amount: T::Balance::default(),
-		})
-	}
-
-	fn increase(&mut self, tranche_tokens_amount: T::Balance) -> DispatchResult {
-		Ok(self
-			.pending_tranche_tokens
-			.ensure_add_assign(tranche_tokens_amount)?)
-	}
-
-	fn decrease(&mut self, tranche_tokens_amount: T::Balance) -> DispatchResult {
-		Ok(self
-			.pending_tranche_tokens
-			.ensure_sub_assign(tranche_tokens_amount)
-			.map_err(|_| Error::<T>::TooMuchDecrease)?)
-	}
-
-	fn pre_swap(
-		&mut self,
-		investment_id: T::InvestmentId,
-		collected: CollectedAmount<T::Balance>,
-	) -> Result<Swap<T::Balance, T::CurrencyId>, DispatchError> {
-		self.base.collected.increase(&collected)?;
-		Ok(Swap {
-			currency_in: self.base.foreign_currency,
-			currency_out: pool_currency_of::<T>(investment_id)?,
-			amount_in: collected.amount_collected,
-		})
-	}
-
-	fn post_swap(
-		&mut self,
-		who: &T::AccountId,
-		investment_id: T::InvestmentId,
-		swapped_amount: T::Balance,
-		pending_amount: T::Balance,
-	) -> Result<Option<ExecutedForeignCollect<T::Balance, T::CurrencyId>>, DispatchError> {
-		self.swapped_amount.ensure_add_assign(swapped_amount)?;
-		if pending_amount.is_zero() {
-			let redemption = T::Investment::redemption(&who, investment_id)?;
-
-			let msg = ExecutedForeignCollect {
-				currency: self.base.foreign_currency,
-				amount_currency_payout: self.swapped_amount,
-				amount_tranche_tokens_payout: self.collected_tranche_tokens(),
-				amount_remaining: redemption,
-			};
-
-			self.pending_tranche_tokens
-				.ensure_sub_assign(self.collected_tranche_tokens())?;
-
-			self.base.collected = CollectedAmount::default();
-			self.swapped_amount = T::Balance::default();
-
-			return Ok(Some(msg));
-		}
-
-		Ok(None)
-	}
-
-	fn is_completed(&self) -> bool {
-		self.pending_tranche_tokens.is_zero()
-	}
-
-	fn collected_tranche_tokens(&self) -> T::Balance {
-		self.base.collected.amount_payment
-	}
-}
-
 pub type SwapOf<T> = Swap<<T as Config>::Balance, <T as Config>::CurrencyId>;
-
-fn pool_currency_of<T: Config>(
-	investment_id: T::InvestmentId,
-) -> Result<T::CurrencyId, DispatchError> {
-	T::PoolInspect::currency_for(investment_id.of_pool()).ok_or(Error::<T>::PoolNotFound.into())
-}
 
 #[frame_support::pallet]
 pub mod pallet {
 	use cfg_traits::{
-		investments::{ForeignInvestment, Investment, InvestmentCollector, TrancheCurrency},
+		investments::{Investment, InvestmentCollector, TrancheCurrency},
 		IdentityCurrencyConversion, PoolInspect, StatusNotificationHook, TokenSwaps,
 	};
-	use cfg_types::investments::{
-		CollectedAmount, ExecutedForeignCollect, ExecutedForeignDecreaseInvest,
-	};
+	use cfg_types::investments::{ExecutedForeignCollect, ExecutedForeignDecreaseInvest};
 	use frame_support::{dispatch::HasCompact, pallet_prelude::*};
-	use sp_runtime::{
-		traits::{AtLeast32BitUnsigned, EnsureAdd, EnsureSub, One, Zero},
-		FixedPointOperand,
-	};
-	use sp_std::cmp::Ordering;
+	use sp_runtime::{traits::AtLeast32BitUnsigned, FixedPointOperand};
 
 	use super::*;
 
@@ -467,7 +157,7 @@ pub mod pallet {
 			CurrencyId = Self::CurrencyId,
 			Balance = Self::Balance,
 			OrderId = Self::SwapId,
-			OrderDetails = Swap<Self::Balance, Self::CurrencyId>,
+			OrderDetails = SwapOf<Self>,
 			SellRatio = Self::BalanceRatio,
 		>;
 
@@ -526,7 +216,7 @@ pub mod pallet {
 		T::AccountId,
 		Blake2_128Concat,
 		T::InvestmentId,
-		InvestmentInfo<T>,
+		entities::InvestmentInfo<T>,
 	>;
 
 	/// Contains the information about the foreign redemption process
@@ -540,7 +230,7 @@ pub mod pallet {
 		T::AccountId,
 		Blake2_128Concat,
 		T::InvestmentId,
-		RedemptionInfo<T>,
+		entities::RedemptionInfo<T>,
 	>;
 
 	/// Maps a `SwapId` to its corresponding `ForeignId`
@@ -576,527 +266,5 @@ pub mod pallet {
 
 		/// The decrease is greater than the current investment/redemption
 		TooMuchDecrease,
-	}
-
-	/// Internal type used as result of `Pallet::apply_swap()`
-	/// Amounts are donominated referenced by the `new_swap` paramenter given to
-	/// `apply_swap()`
-	#[derive(Debug, PartialEq)]
-	pub(crate) struct SwapStatus<T: Config> {
-		/// The amount (in) already swapped and available to use.
-		pub swapped: T::Balance,
-
-		/// The amount (in) pending to be swapped
-		pub pending: T::Balance,
-
-		/// The swap id for a possible reminder swap order after `apply_swap()`
-		pub swap_id: Option<T::SwapId>,
-	}
-
-	impl<T: Config> Pallet<T> {
-		/// Inserts, updates or removes a swap id associated to a foreign
-		/// action.
-		fn update_swap_id(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			action: Action,
-			new_swap_id: Option<T::SwapId>,
-		) -> DispatchResult {
-			let previous_swap_id = ForeignIdToSwapId::<T>::get((who, investment_id, action));
-
-			if previous_swap_id != new_swap_id {
-				if let Some(new_id) = new_swap_id {
-					SwapIdToForeignId::<T>::insert(new_id, (who.clone(), investment_id, action));
-					ForeignIdToSwapId::<T>::insert((who.clone(), investment_id, action), new_id);
-				}
-
-				if let Some(old_id) = previous_swap_id {
-					SwapIdToForeignId::<T>::remove(old_id);
-					ForeignIdToSwapId::<T>::remove((who.clone(), investment_id, action));
-				}
-			}
-
-			Ok(())
-		}
-
-		pub fn pending_swap_amount(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			currency_in: T::CurrencyId,
-			action: Action,
-		) -> T::Balance {
-			ForeignIdToSwapId::<T>::get((who, investment_id, action))
-				.map(|swap_id| T::TokenSwaps::get_order_details(swap_id))
-				.flatten()
-				.filter(|swap| swap.currency_in == currency_in)
-				.map(|swap| swap.amount_in)
-				.unwrap_or(T::Balance::default())
-		}
-
-		/// Apply a swap over a current possible swap state.
-		/// - If there was no previous swap, it adds it.
-		/// - If there was a swap in the same direction, it increments it.
-		/// - If there was a swap in the opposite direction:
-		///   - If the amount is smaller, it decrements it.
-		///   - If the amount is the same, it removes the inverse swap.
-		///   - If the amount is greater, it removes the inverse swap and create
-		///     another with the excess
-		///
-		/// The returned status contains the swapped amounts after this call and
-		/// the pending amounts to be swapped of both swap directions.
-		pub(crate) fn apply_swap(
-			who: &T::AccountId,
-			new_swap: SwapOf<T>,
-			over_swap_id: Option<T::SwapId>,
-		) -> Result<SwapStatus<T>, DispatchError> {
-			match over_swap_id {
-				None => {
-					let swap_id = T::TokenSwaps::place_order(
-						who.clone(),
-						new_swap.currency_in,
-						new_swap.currency_out,
-						new_swap.amount_in,
-						T::BalanceRatio::one(),
-					)?;
-
-					Ok(SwapStatus {
-						swapped: T::Balance::zero(),
-						pending: new_swap.amount_in,
-						swap_id: Some(swap_id),
-					})
-				}
-				Some(swap_id) => {
-					let swap = T::TokenSwaps::get_order_details(swap_id)
-						.ok_or(Error::<T>::SwapOrderNotFound)?;
-
-					if swap.is_same_direction(&new_swap)? {
-						let amount_to_swap = swap.amount_in.ensure_add(new_swap.amount_in)?;
-						T::TokenSwaps::update_order(
-							who.clone(),
-							swap_id,
-							amount_to_swap,
-							T::BalanceRatio::one(),
-						)?;
-
-						Ok(SwapStatus {
-							swapped: T::Balance::zero(),
-							pending: amount_to_swap,
-							swap_id: Some(swap_id),
-						})
-					} else {
-						let inverse_swap = swap;
-
-						let new_swap_amount_out = T::CurrencyConverter::stable_to_stable(
-							new_swap.currency_out,
-							new_swap.currency_in,
-							new_swap.amount_in,
-						)?;
-
-						match inverse_swap.amount_in.cmp(&new_swap_amount_out) {
-							Ordering::Greater => {
-								let amount_to_swap =
-									inverse_swap.amount_in.ensure_sub(new_swap_amount_out)?;
-
-								T::TokenSwaps::update_order(
-									who.clone(),
-									swap_id,
-									amount_to_swap,
-									T::BalanceRatio::one(),
-								)?;
-
-								Ok(SwapStatus {
-									swapped: new_swap.amount_in,
-									pending: T::Balance::zero(),
-									swap_id: Some(swap_id),
-								})
-							}
-							Ordering::Equal => {
-								T::TokenSwaps::cancel_order(swap_id)?;
-
-								Ok(SwapStatus {
-									swapped: new_swap.amount_in,
-									pending: T::Balance::zero(),
-									swap_id: None,
-								})
-							}
-							Ordering::Less => {
-								T::TokenSwaps::cancel_order(swap_id)?;
-
-								let inverse_swap_amount_out =
-									T::CurrencyConverter::stable_to_stable(
-										inverse_swap.currency_out,
-										inverse_swap.currency_in,
-										inverse_swap.amount_in,
-									)?;
-
-								let amount_to_swap =
-									new_swap.amount_in.ensure_sub(inverse_swap_amount_out)?;
-
-								let swap_id = T::TokenSwaps::place_order(
-									who.clone(),
-									new_swap.currency_in,
-									new_swap.currency_out,
-									amount_to_swap,
-									T::BalanceRatio::one(),
-								)?;
-
-								Ok(SwapStatus {
-									swapped: inverse_swap_amount_out,
-									pending: amount_to_swap,
-									swap_id: Some(swap_id),
-								})
-							}
-						}
-					}
-				}
-			}
-		}
-
-		/// A wrap over `apply_swap()` that takes care of updating the swap id
-		/// and notify
-		fn apply_swap_and_notify(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			action: Action,
-			new_swap: SwapOf<T>,
-		) -> DispatchResult {
-			let swap_id = ForeignIdToSwapId::<T>::get((who, investment_id, action));
-
-			let status = Self::apply_swap(who, new_swap.clone(), swap_id)?;
-
-			Self::update_swap_id(who, investment_id, action, status.swap_id)?;
-
-			if !status.swapped.is_zero() {
-				Self::notify_swap_done(
-					who,
-					investment_id,
-					action,
-					new_swap.currency_out,
-					status.swapped,
-					status.pending,
-				)?;
-			}
-
-			Ok(())
-		}
-
-		/// Notifies that a partial swap has been done and applies the result to
-		/// an `InvestmentInfo` or `RedemptionInfo`
-		fn notify_swap_done(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			action: Action,
-			currency_out: T::CurrencyId,
-			swapped_amount: T::Balance,
-			pending_amount: T::Balance,
-		) -> DispatchResult {
-			match action {
-				Action::Investment => Pallet::<T>::notify_investment_swap_done(
-					&who,
-					investment_id,
-					currency_out,
-					swapped_amount,
-					pending_amount,
-				),
-				Action::Redemption => Pallet::<T>::notify_redemption_swap_done(
-					&who,
-					investment_id,
-					swapped_amount,
-					pending_amount,
-				),
-			}
-		}
-
-		/// Notifies that a partial swap has been done and applies the result to
-		/// an `InvestmentInfo`
-		fn notify_investment_swap_done(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			currency_out: T::CurrencyId,
-			swapped: T::Balance,
-			pending: T::Balance,
-		) -> DispatchResult {
-			let msg = ForeignInvestmentInfo::<T>::mutate_exists(&who, investment_id, |entry| {
-				let info = entry.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-
-				if currency_out == info.base.foreign_currency {
-					info.post_increase_swap(who, investment_id, swapped)?;
-
-					Ok(None)
-				} else {
-					let msg = info.post_decrease_swap(investment_id, swapped, pending)?;
-
-					if info.is_completed()? {
-						*entry = None;
-					}
-
-					Ok::<_, DispatchError>(msg)
-				}
-			})?;
-
-			// We send the event out of the Info mutation closure
-			if let Some(msg) = msg {
-				T::DecreasedForeignInvestOrderHook::notify_status_change(
-					(who.clone(), investment_id),
-					msg,
-				)?;
-			}
-
-			Ok(())
-		}
-
-		/// Notifies that a partial swap has been done and applies the result to
-		/// an `RedemptionInfo`
-		fn notify_redemption_swap_done(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			swapped_amount: T::Balance,
-			pending_amount: T::Balance,
-		) -> DispatchResult {
-			let msg = ForeignRedemptionInfo::<T>::mutate_exists(&who, investment_id, |entry| {
-				let info = entry.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				let msg = info.post_swap(who, investment_id, swapped_amount, pending_amount)?;
-
-				if info.is_completed() {
-					*entry = None;
-				}
-
-				Ok::<_, DispatchError>(msg)
-			})?;
-
-			// We send the event out of the Info mutation closure
-			if let Some(msg) = msg {
-				T::CollectedForeignRedemptionHook::notify_status_change(
-					(who.clone(), investment_id),
-					msg,
-				)?;
-			}
-
-			Ok(())
-		}
-	}
-
-	impl<T: Config> ForeignInvestment<T::AccountId> for Pallet<T> {
-		type Amount = T::Balance;
-		type CurrencyId = T::CurrencyId;
-		type Error = DispatchError;
-		type InvestmentId = T::InvestmentId;
-
-		fn increase_foreign_investment(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			foreign_amount: T::Balance,
-			foreign_currency: T::CurrencyId,
-		) -> DispatchResult {
-			let swap = ForeignInvestmentInfo::<T>::mutate(&who, investment_id, |info| {
-				let info = info.get_or_insert(InvestmentInfo::new(foreign_currency)?);
-				info.base.ensure_same_foreign(foreign_currency)?;
-				info.pre_increase_swap(investment_id, foreign_amount)
-			})?;
-
-			Self::apply_swap_and_notify(who, investment_id, Action::Investment, swap)
-		}
-
-		fn decrease_foreign_investment(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			foreign_amount: T::Balance,
-			foreign_currency: T::CurrencyId,
-		) -> DispatchResult {
-			let swap = ForeignInvestmentInfo::<T>::mutate(&who, investment_id, |info| {
-				let info = info.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				info.base.ensure_same_foreign(foreign_currency)?;
-				info.pre_decrease_swap(who, investment_id, foreign_amount)
-			})?;
-
-			Self::apply_swap_and_notify(who, investment_id, Action::Investment, swap)
-		}
-
-		fn increase_foreign_redemption(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			tranche_tokens_amount: T::Balance,
-			payout_foreign_currency: T::CurrencyId,
-		) -> DispatchResult {
-			ForeignRedemptionInfo::<T>::mutate(who, investment_id, |info| -> DispatchResult {
-				let info = info.get_or_insert(RedemptionInfo::new(payout_foreign_currency)?);
-				info.base.ensure_same_foreign(payout_foreign_currency)?;
-				info.increase(tranche_tokens_amount)
-			})?;
-
-			T::Investment::update_redemption(
-				who,
-				investment_id,
-				T::Investment::redemption(who, investment_id)?.ensure_add(tranche_tokens_amount)?,
-			)
-		}
-
-		fn decrease_foreign_redemption(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			tranche_tokens_amount: T::Balance,
-			payout_foreign_currency: T::CurrencyId,
-		) -> DispatchResult {
-			ForeignRedemptionInfo::<T>::mutate(&who, investment_id, |info| {
-				let info = info.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				info.base.ensure_same_foreign(payout_foreign_currency)?;
-				info.decrease(tranche_tokens_amount)
-			})?;
-
-			T::Investment::update_redemption(
-				who,
-				investment_id,
-				T::Investment::redemption(who, investment_id)?.ensure_sub(tranche_tokens_amount)?,
-			)
-		}
-
-		fn collect_foreign_investment(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			payment_foreign_currency: T::CurrencyId,
-		) -> DispatchResult {
-			ForeignInvestmentInfo::<T>::mutate(&who, investment_id, |info| {
-				let info = info.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				info.base.ensure_same_foreign(payment_foreign_currency)?;
-				Ok::<_, DispatchError>(())
-			})?;
-
-			T::Investment::collect_investment(who.clone(), investment_id)
-		}
-
-		fn collect_foreign_redemption(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-			payout_foreign_currency: T::CurrencyId,
-			_pool_currency: T::CurrencyId,
-		) -> DispatchResult {
-			ForeignRedemptionInfo::<T>::mutate(&who, investment_id, |info| {
-				let info = info.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				info.base.ensure_same_foreign(payout_foreign_currency)?;
-				Ok::<_, DispatchError>(())
-			})?;
-
-			T::Investment::collect_redemption(who.clone(), investment_id)
-		}
-
-		fn investment(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-		) -> Result<T::Balance, DispatchError> {
-			T::Investment::investment(who, investment_id)
-		}
-
-		fn redemption(
-			who: &T::AccountId,
-			investment_id: T::InvestmentId,
-		) -> Result<T::Balance, DispatchError> {
-			T::Investment::redemption(who, investment_id)
-		}
-
-		fn accepted_payment_currency(
-			investment_id: T::InvestmentId,
-			currency: T::CurrencyId,
-		) -> bool {
-			if T::Investment::accepted_payment_currency(investment_id, currency) {
-				true
-			} else {
-				T::PoolInspect::currency_for(investment_id.of_pool())
-					.map(|pool_currency| T::TokenSwaps::valid_pair(pool_currency, currency))
-					.unwrap_or(false)
-			}
-		}
-
-		fn accepted_payout_currency(
-			investment_id: T::InvestmentId,
-			currency: T::CurrencyId,
-		) -> bool {
-			if T::Investment::accepted_payout_currency(investment_id, currency) {
-				true
-			} else {
-				T::PoolInspect::currency_for(investment_id.of_pool())
-					.map(|pool_currency| T::TokenSwaps::valid_pair(currency, pool_currency))
-					.unwrap_or(false)
-			}
-		}
-	}
-
-	pub struct FulfilledSwapOrderHook<T>(PhantomData<T>);
-	impl<T: Config> StatusNotificationHook for FulfilledSwapOrderHook<T> {
-		type Error = DispatchError;
-		type Id = T::SwapId;
-		type Status = SwapOf<T>;
-
-		fn notify_status_change(
-			swap_id: T::SwapId,
-			last_swap: SwapOf<T>,
-		) -> Result<(), DispatchError> {
-			let (who, investment_id, action) =
-				SwapIdToForeignId::<T>::get(swap_id).ok_or(Error::<T>::SwapOrderNotFound)?;
-
-			let pending_amount = match T::TokenSwaps::get_order_details(swap_id) {
-				Some(swap) => swap.amount_in,
-				None => {
-					Pallet::<T>::update_swap_id(&who, investment_id, action, None)?;
-					T::Balance::default()
-				}
-			};
-
-			Pallet::<T>::notify_swap_done(
-				&who,
-				investment_id,
-				action,
-				last_swap.currency_out,
-				last_swap.amount_in,
-				pending_amount,
-			)
-		}
-	}
-
-	pub struct CollectedInvestmentHook<T>(PhantomData<T>);
-	impl<T: Config> StatusNotificationHook for CollectedInvestmentHook<T> {
-		type Error = DispatchError;
-		type Id = (T::AccountId, T::InvestmentId);
-		type Status = CollectedAmount<T::Balance>;
-
-		fn notify_status_change(
-			(who, investment_id): (T::AccountId, T::InvestmentId),
-			collected: CollectedAmount<T::Balance>,
-		) -> DispatchResult {
-			let msg = ForeignInvestmentInfo::<T>::mutate_exists(&who, investment_id, |entry| {
-				let info = entry.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				let msg = info.post_collect(investment_id, collected);
-
-				if info.is_completed()? {
-					*entry = None;
-				}
-
-				msg
-			})?;
-
-			// We send the event out of the Info mutation closure
-			T::CollectedForeignInvestmentHook::notify_status_change(
-				(who.clone(), investment_id),
-				msg,
-			)
-		}
-	}
-
-	pub struct CollectedRedemptionHook<T>(PhantomData<T>);
-	impl<T: Config> StatusNotificationHook for CollectedRedemptionHook<T> {
-		type Error = DispatchError;
-		type Id = (T::AccountId, T::InvestmentId);
-		type Status = CollectedAmount<T::Balance>;
-
-		fn notify_status_change(
-			(who, investment_id): (T::AccountId, T::InvestmentId),
-			collected: CollectedAmount<T::Balance>,
-		) -> DispatchResult {
-			let swap = ForeignRedemptionInfo::<T>::mutate(&who, investment_id, |info| {
-				let info = info.as_mut().ok_or(Error::<T>::InfoNotFound)?;
-				info.pre_swap(investment_id, collected)
-			})?;
-
-			Pallet::<T>::apply_swap_and_notify(&who, investment_id, Action::Redemption, swap)
-		}
 	}
 }
